nextjs-fastapi-master/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo
python-jose[cryptography]
passlib[bcrypt]
motor
python-multipart
python-dotenv
bcrypt==4.0.1